+++ changelog.md
@@ -8,6 +8,7 @@
 * [4479](https://github.com/zeta-chain/node/pull/4479) - add MsgRemoveObserver
 * [4484](https://github.com/zeta-chain/node/pull/4484) - change params for 2s block time
 * [4507](https://github.com/zeta-chain/node/pull/4507) - add a tool to list tss balances across all supported chains
+* [4492](https://github.com/zeta-chain/node/pull/4492) - add migration script to set authorization for MsgRemoveObserver
 * [4453](https://github.com/zeta-chain/node/pull/4453) - add an option to create dry zetaclients to connect to live networks
 * [4493](https://github.com/zeta-chain/node/pull/4493) - enable ALT in solana inbounds